backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
sqlalchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.3
pgvector==0.3.2
pydantic==2.9.2
pydantic-settings==2.5.2
google-generativeai==0.8.2
python-dotenv==1.0.1
httpx==0.27.2
pillow==10.4.0
aiofiles==24.1.0
boto3==1.35.36
psycopg2-binary==2.9.9
